Fix Lua function parser helpers after PCH commit

+++ Server/mods/deathmatch/logic/luadefs/CLuaPedDefs.cpp
@@ -179,7 +179,7 @@ int CLuaPedDefs::GetValidPedModels(lua_State* luaVM)
 {
     int iIndex = 0;
     lua_newtable(luaVM);
-    for (int i = 0; i <= 312; i++)
+    for (unsigned short i = 0; i <= 312; i++)
     {
         if (CPlayerManager::IsValidPlayerModel(i))
         {
